Ensure Relational tests are run in Maven build (#205)
