+++ legend-engine-executionPlan-execution-store-relational/src/main/java/org/finos/legend/engine/plan/execution/stores/relational/connection/manager/strategic/DataSourceSpecificationTransformer.java
@@ -101,7 +101,6 @@ public class DataSourceSpecificationTransformer implements DatasourceSpecificati
             }
             else
             {
-                System.out.println("---> "+((StaticDataSourceSpecificationKey) key).getPort());
                 return new StaticDataSourceSpecification(
                         (StaticDataSourceSpecificationKey) key,
                         new H2Manager(),
